RIsolti alcuni bugs di interfaccia e gestito il possesso delle partite

+++ server/utils.c
@@ -184,7 +184,12 @@ cJSON* build_message(int socket_fd, message_type_enum message_type, generic_buff
             cJSON_AddItemToObject(content, "draw_result", handle_draw(&buffer->handle_draw, mem.match_list));
         break;
         case SIG_DELETE_MATCH:
-            cJSON_AddItemToObject(content, "match_list", delete_match(buffer->delete_match.match_id, mem.match_list));
+            if(mem.match_list[buffer->delete_match.match_id].owner_id != socket_fd){
+                cJSON_AddStringToObject(content, "info", "not your match");
+            }
+            else{
+                cJSON_AddItemToObject(content, "match_list", delete_match(buffer->delete_match.match_id, mem.match_list));
+            }
         break;
     }
